# Conflicts:
#	src/main/java/app/model/Jobs.java
#	src/main/java/app/model/ProfessionJob.java
#	src/main/java/app/model/UserRecruiterJobRequirement.java
#	target/classes/app/model/Jobs.class
#	target/classes/app/model/ProfessionJob.class
#	target/classes/app/model/UserRecruiterJobRequirement.class
#	target/classes/app/model/Users.class
#	target/classes/app/model/UsersFreelancer.class
#	target/classes/app/model/UsersRecruiter.class

+++ src/main/java/app/repository/JobsRepository.java
@@ -3,5 +3,5 @@ package app.repository;
 import app.model.Jobs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobsRepository extends JpaRepository<Jobs, Long> {
+public interface JobsRepository extends JpaRepository<Jobs, Integer> {
 }
