+++ src/main/java/app/repository/ProfessionJobRepository.java
@@ -3,5 +3,5 @@ package app.repository;
 import app.model.ProfessionJob;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfessionJobRepository extends JpaRepository<ProfessionJob,Long> {
+public interface ProfessionJobRepository extends JpaRepository<ProfessionJob,Integer> {
 }
